Update isOpenLoop
Updated isOpenLoop in swerveSubsystem to `true` to make autonomous work.  If isOpenLoop is set to false, SwerveModule.java will attempt to use a set velocity to drive instead of motor percent, which caused the issues we saw in autonomous with the rapid square wave oscillation.  Teleop is also much smoother now, as the robot doesn't keep trying to speed check itself if you go too fast, causing jerking.