ArmMed
Added an ArmMed command. Created and zeroed the arm positions needed for ArmMed. Also added a bit of comments on the constants page for the Arm values, nothing major.

Still need:
Comments on the three arm positions
To account for gearbox ratio (unless that has already been solved)
Medium positions for the arm motors
Fine tuning the PID and position values